+++ src/app/form-fields/form-fields.component.ts
@@ -5,7 +5,7 @@ import {
   Component,
   Input,
   Output,
-  EventEmitter, OnChanges, SimpleChanges
+  EventEmitter, OnChanges, AfterViewInit, SimpleChanges, ViewChild
 } from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {FetchService} from '../services/fetch.service';
@@ -14,20 +14,21 @@ import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 import {FormService} from '../services/form.service';
 import { fhir } from '../fhir';
 import {Util} from '../lib/util';
+import {FormComponent, FormProperty} from '@lhncbc/ngx-schema-form';
 
 @Component({
   selector: 'lfb-form-fields',
   template: `
     <div class="card-body content">
-      <div  *ngIf="true">
+      <div>
         <h4 class="ml-2">Form level attributes</h4>
         <p class="ml-4">Enter basic information about the form.</p>
         <hr/>
         <div class="container">
-          <sf-form [schema]="qlSchema"
+          <sf-form #ngxForm [schema]="qlSchema"
                    [model]="questionnaire"
                    (onChange)="valueChanged($event)"
-                   (modelReset)="onFormFieldsLoaded()"
+                   (modelReset)="onFormFieldsLoaded($event)"
           ></sf-form>
         </div>
         <hr/>
@@ -44,8 +45,9 @@ import {Util} from '../lib/util';
     }
   `]
 })
-export class FormFieldsComponent implements OnChanges {
+export class FormFieldsComponent implements OnChanges, AfterViewInit {
 
+  @ViewChild('ngxForm', {static: false, read: FormComponent}) ngxForm!: FormComponent;
   @Input()
   questionsButtonLabel = 'Create questions';
   @Input()
@@ -61,6 +63,7 @@ export class FormFieldsComponent implements OnChanges {
   @Output()
   questionnaireChange = new EventEmitter<fhir.Questionnaire>();
   loading = false;
+  formValue: fhir.Questionnaire;
   constructor(
     private http: HttpClient,
     private dataSrv: FetchService,
@@ -70,6 +73,9 @@ export class FormFieldsComponent implements OnChanges {
     this.qlSchema = this.formService.getFormLevelSchema();
   }
 
+  ngAfterViewInit() {
+    this.adjustRootFormProperty();
+  }
 
   ngOnChanges(changes: SimpleChanges) {
     if(changes.questionnaire) {
@@ -77,8 +83,25 @@ export class FormFieldsComponent implements OnChanges {
     }
   }
 
-  onFormFieldsLoaded() {
+  adjustRootFormProperty(): boolean {
+    let ret = false;
+    const rootProperty = this.ngxForm?.rootProperty;
+    // Emit the value after any adjustments.
+    // Set '__$codeYesNo' to true, when 'code' is present. The default is false.
+    if(!Util.isEmpty(rootProperty?.searchProperty('/code').value)) {
+      // Loading is done. Change of value should emit the value in valueChanged().
+      rootProperty?.searchProperty('/__$codeYesNo').setValue(true, false);
+      ret = true;
+    }
+    return ret;
+  }
+
+  onFormFieldsLoaded(event) {
     this.loading = false;
+    this.formValue = event.value;
+    if(!this.adjustRootFormProperty()) {
+      this.questionnaireChange.emit(Util.convertToQuestionnaireJSON(event.value));
+    }
   }
   /**
    * Send message to base page to switch the view.