+++ src/app/lib/widgets/answer-option-methods/answer-option-methods.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { AnswerOptionMethodsComponent } from './answer-option-methods.component';
+
+describe('LfbAnswerOptionComponent', () => {
+  let component: AnswerOptionMethodsComponent;
+  let fixture: ComponentFixture<AnswerOptionMethodsComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ AnswerOptionMethodsComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(AnswerOptionMethodsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  xit('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});